+++ lxcraft/debian/apt_package.py
@@ -1,6 +1,5 @@
 import os
 from dataclasses import dataclass
-from functools import partial
 from subprocess import getstatusoutput
 
 import lxcraft
@@ -13,16 +12,16 @@ class AptPackages(lxcraft.PlanElement):
     package_list: list[str]
 
     def get_actions(self):
-        package_list = []
+        self.missing_package_list = []
         for package in self.package_list:
             if self.is_installed(package):
                 continue
-            package_list.append(package)
-        if not package_list:
-            return
-        return [partial(self.package_action, " ".join(package_list))]
+            self.missing_package_list.append(package)
+        if self.missing_package_list:
+            return self.install_missing()
 
-    def package_action(self, package_list: str):
+    def install_missing(self):
+        package_list = " ".join(self.missing_package_list)
         rc = os.system(f"apt install -y {package_list}")
         if rc != 0:
             raise Exception(f"Command terminated with non zero exit code {rc}")
@@ -31,3 +30,6 @@ class AptPackages(lxcraft.PlanElement):
     def is_installed(package_name: str):
         rc, _ = getstatusoutput(f"dpkg -s {package_name}")
         return rc == 0
+
+    def destroy(self):
+        os.system(f"apt remove -y {' '.join(self.package_list)}")