+++ tests/test_directory.py
@@ -7,9 +7,6 @@ DIRNAME = "/run/user/www-data"
 
 
 def test_directory_create():
-    Plan([Directory(DIRNAME)]).execute()
-    assert Path(DIRNAME).is_dir()
-
-
-def test_user_mutiple():
-    test_directory_create()
+    with Plan([Directory(DIRNAME)]) as plan:
+        plan.execute()
+        assert Path(DIRNAME).is_dir()