+++ tests/test_user.py
@@ -1,5 +1,7 @@
 import pwd
 
+import pytest
+
 from lxcraft import Plan
 from lxcraft.user import User
 
@@ -7,5 +9,24 @@ USERNAME = "lxcraft"
 
 
 def test_user_add():
-    Plan(User(USERNAME)).execute()
-    assert pwd.getpwnam(USERNAME)
+    with Plan(User(USERNAME)) as plan:
+        plan.execute()
+        assert pwd.getpwnam(USERNAME)
+
+    with Plan(
+        User(
+            USERNAME,
+            "geocs",
+            "password",
+            "/bin/bash",
+            f"/home/{USERNAME}",
+            2000,
+            1,
+            ["root"],
+            True,
+        )
+    ) as plan:
+        plan.execute()
+        assert pwd.getpwnam(USERNAME)
+    with pytest.raises(Exception, match=r"Command terminated with non zero exit code"):
+        Plan(User("")).execute()