+++ CHANGELOG.md
@@ -1,15 +1,16 @@
 # CHANGELOG.md
 
 
-## 0.0.3 (UNRELEASED)
+## 0.0.3 (26 July 2023)
 
 Improvements:
 
-  - Added coverage report (100% on this version)
+  - Added coverage report
   - Plan:
     - removed 'name' attribute
-    - added `Resource` class
     - added one or list of resources
     - rename .run() to .execute()
+    - Renamed `Element` to `Resource`
+    - Documented the Resource API at `docs/resource.md`
   - FileContent:
     - added support for search in the templates dir